split riverdrivertest.go
Instead of a giant 5000 line file, split into smaller files grouped by
entity. For job queries where we have many tests, break them into files
by query type.
